requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
scipy
joblib
pydantic
PyYAML